Se agrega funcionalidad al componente de SendReview y se agrega hook para dicha funcionalidad
